Move *Io to separate files. Fix lint/mypy
